Get convars at the start instead of making them static

+++ src/features/ticks/ticks.h
@@ -5,7 +5,7 @@
 namespace TickManager
 {
 	extern bool m_bSendPacket;
-	extern int m_iChokedCommands;
+	extern uint8_t m_iChokedCommands;
 
 	void CL_Move(float accumulated_extra_samples, bool bFinalTick);
 	void CL_SendMove(void);
